Refactor Settings Menu Element into seperate file
